requirements.txt
requests>=2.25.1
beautifulsoup4>=4.9.3
pyOpenSSL>=21.0.0
python-dotenv>=0.19.0
Pillow>=8.3.0
playwright>=1.40.0